Rename 'GetRandomInt' function to 'GetRandomIntFromRange', and fix related import in Instance class
